+++ src/tcp_local_lat.c
@@ -37,12 +37,12 @@
 #include <unistd.h>
 
 int main(int argc, char *argv[]) {
-  int size;
+  ssize_t size;
   char *buf;
   int64_t count, i;
 
   ssize_t len;
-  size_t sofar;
+  ssize_t sofar;
 
   int yes = 1;
   int ret;
@@ -67,7 +67,7 @@ int main(int argc, char *argv[]) {
     return 1;
   }
 
-  printf("message size: %i octets\n", size);
+  printf("message size: %li octets\n", size);
   printf("roundtrip count: %li\n", count);
 
   memset(&hints, 0, sizeof hints);
